services/ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
structlog==24.4.0
httpx==0.28.1
# pytesseract==0.3.13  # TODO: add when tesseract-ocr is installed
pdfplumber==0.11.4
Pillow==11.1.0
python-multipart==0.0.20
